update manifest endpoint is 90% integrated

+++ DeliveryManager.Server/Models/deliveryForm.cs
@@ -10,7 +10,7 @@ namespace DeliveryManager.Server.Models
         public string LASTUPDATE { get; set; } = string.Empty;
         public string MFSTNUMBER { get; set; } = string.Empty;
         public string POWERUNIT { get; set; } = string.Empty;
-        public int STOP { get; set; } = 0;
+        public short STOP { get; set; } = 0;
         public string MFSTDATE { get; set; } = string.Empty;
         public string PRONUMBER { get; set; } = string.Empty;
         public string PRODATE { get; set; } = string.Empty;
@@ -21,12 +21,12 @@ namespace DeliveryManager.Server.Models
         public string CONSCITY { get; set; } = string.Empty;
         public string CONSSTATE { get; set; } = string.Empty;
         public string CONSZIP { get; set; } = string.Empty;
-        public int TTLPCS { get; set; } = 0;
-        public int TTLYDS { get; set; } = 0;
-        public int TTLWGT { get; set; } = 0;
+        public short TTLPCS { get; set; } = 0;
+        public short TTLYDS { get; set; } = 0;
+        public short TTLWGT { get; set; } = 0;
         public string? DLVDDATE { get; set; }
         public string? DLVDTIME { get; set; }
-        public int? DLVDPCS { get; set; }
+        public short? DLVDPCS { get; set; }
         public string? DLVDSIGN { get; set; }
         public string? DLVDNOTE { get; set; }
         public IFormFile? DLVDIMGFILELOCN { get; set; }
